+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 - Add function to chmod files/folders transferred via Globus (#1739)
 - Increase wall clock time to 14 days for dori site (#1754))
 - change NMDC site's refdata dir path (#1756)
+- add function to chmod files/folders transferred via Globus (#1739)
+- try to address db update failures after to long transfer operation (#1746)
 
 ### External Facing Changes